+++ pkg/datarepair/checker/checker_test.go
@@ -1,4 +1,15 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package checker
+package datarepair
+
+import (
+	"testing"
+)
+
+func TestIdentifyInjuredSegments(t *testing.T) {
+}
+
+func BenchmarkIdentifyInjuredSegments(b *testing.B) {
+
+}
