--- pkg/datarepair/checker_test.go
@@ -1,15 +1,0 @@
-// Copyright (C) 2018 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-package datarepair
-
-import (
-	"testing"
-)
-
-func TestIdentifyInjuredSegments(t *testing.T) {
-}
-
-func BenchmarkIdentifyInjuredSegments(b *testing.B) {
-
-}
